requirements.txt
spotipy==2.20.0
pylast==5.1.0
python-dateutil==2.8.2
